Problem: when bound to any adapter "*" zyre will publish it's return connection endpoint in the HELLO message as a hostname, however a condition exists where hostname lookup can fail or return a different route to the original packet such as on a different subnet (perhaps over wifi where the beacon came via ethernet).
Solution: In the case of gossip discovery the endpoint must be set by the zyre caller, but otherwise zyre should never send a fixed endpoint when not using gossip as the endpoint may change with network changes or between connections from different source networks. In these cases the node should use the remote endpoint to find a suitable local interface and it's public ip address to inform the remote peer to connect back to us on. If this is not possible (no ip address received in HELLO for the remote peer) the node will resort to sending it's hostname as previous